transformers/appairflow4/requirements.txt
azure-storage-file-share
azure-storage-blob
pandas
pyyaml
pyarrow
